requirements.txt
click
frictionless
pandas
pyld
pyyaml
rdflib
requests
